+++ dart/dynamics/CMakeLists.txt
@@ -18,12 +18,17 @@ dart_generate_include_header_file(
 
 # Install
 install(
-  FILES ${hdrs} ${CMAKE_CURRENT_BINARY_DIR}/dynamics.hpp
+  FILES ${hdrs} ikfast.h ${CMAKE_CURRENT_BINARY_DIR}/dynamics.hpp
   DESTINATION include/dart/dynamics
   COMPONENT headers
 )
 install(
   FILES ${detail_hdrs}
   DESTINATION include/dart/dynamics/detail
+  COMPONENT headers
+)
+install(
+  FILES ikfast.h
+  DESTINATION include/dart/external/ikfast
   COMPONENT headers
 )
