+++ dart/dynamics/IkFast.hpp
@@ -36,9 +36,8 @@
 #include <array>
 
 #define IKFAST_HAS_LIBRARY
-#include "dart/external/ikfast/ikfast.h"
-
 #include <dart/dynamics/InverseKinematics.hpp>
+#include <dart/dynamics/ikfast.h>
 
 namespace dart {
 namespace dynamics {
