+++ dart/external/CMakeLists.txt
@@ -2,6 +2,5 @@ add_subdirectory(convhull_3d)
 if(NOT DART_USE_SYSTEM_IMGUI)
   add_subdirectory(imgui)
 endif()
-add_subdirectory(ikfast)
 add_subdirectory(lodepng)
 add_subdirectory(odelcpsolver)
